Movimento mais sensível do player

+++ src/com/zelda/entities/Entities.java
@@ -5,8 +5,8 @@ import java.awt.image.BufferedImage;
 
 public class Entities {
 	
-	protected int x;
-	protected int y;
+	protected double x;
+	protected double y;
 	protected int whidth;
 	protected int height;
 	
@@ -28,11 +28,11 @@ public class Entities {
 	}
 
 	public int getX() {
-		return x;
+		return (int)x;
 	}
 
 	public int getY() {
-		return y;
+		return (int)y;
 	}
 
 	public int getWhidth() {
